A finalized implementation of the priority queue
We sort in reverse now and we have a dedicated resort method
instead of the repeated code in enqueue and dequeue. This comes
at a cost of the user knowing when to resort the queue (instead
this automatically happening on every enqueue and dequeue).

The automatic approach doesn't work because someone would always
dequeue an item and change it afterwards. The sorting it right after
the  dequeue operation would miss on key changes after the operation.